requirements.txt
appdirs==1.4.4
asttokens==2.4.1
atomicwrites==1.4.0
chardet==3.0.4
colorama==0.4.6
comm==0.2.2
debugpy==1.8.6
decorator==5.1.1
emoji==2.14.0
executing==2.1.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.10.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jsonpointer==2.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mkl-service==2.4.0
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==3.20.3
psutil==6.0.0
pure_eval==0.2.3
pyasn1-modules==0.2.8
pycurl==7.45.2
Pygments==2.18.0
pyls-spyder==0.4.0
PyQt5==5.15.10
PyQtWebEngine==5.15.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
pywin32==306
pyzmq==26.2.0
rfc3986==1.5.0
setuptools==69.5.1
six==1.16.0
stack-data==0.6.3
textblob==0.18.0.post0
tornado==6.4.1
traitlets==5.14.3
wcwidth==0.2.13
wheel==0.43.0
